# docs/manager/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "manager service API documentation",
        "title": "manager service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/intents": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete one or more schedule intents.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Strategies"
                ],
                "summary": "Delete schedule intents",
                "parameters": [
                    {
                        "description": "Intent IDs to delete",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.DeleteScheduleIntentsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.SuccessResponse-github_com_Gthulhu_api_manager_rest_EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/intents/self": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List schedule intents created by the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Strategies"
                ],
                "summary": "List self schedule intents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.SuccessResponse-rest_ListScheduleIntentsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/nodes/{nodeID}/pods/pids": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all pods running on the specified node with their associated process IDs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nodes"
                ],
                "summary": "Get Pod-PID mapping for a specific node",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node ID",
                        "name": "nodeID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.SuccessResponse-rest_GetNodePodPIDMappingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/strategies": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new schedule strategy.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Strategies"
                ],
                "summary": "Create schedule strategy",
                "parameters": [
                    {
                        "description": "Schedule strategy payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.CreateScheduleStrategyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.SuccessResponse-github_com_Gthulhu_api_manager_rest_EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a schedule strategy and its associated intents.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Strategies"
                ],
                "summary": "Delete schedule strategy",
                "parameters": [
                    {
                        "description": "Strategy ID to delete",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.DeleteScheduleStrategyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.SuccessResponse-github_com_Gthulhu_api_manager_rest_EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/strategies/self": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List schedule strategies created by the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Strategies"
                ],
                "summary": "List self schedule strategies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.SuccessResponse-rest_ListSchedulerStrategiesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_Gthulhu_api_manager_rest.EmptyResponse": {
            "type": "object"
        },
        "github_com_Gthulhu_api_manager_rest.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "github_com_Gthulhu_api_manager_rest.LabelSelector": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "github_com_Gthulhu_api_manager_rest.SuccessResponse-github_com_Gthulhu_api_manager_rest_EmptyResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.EmptyResponse"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "github_com_Gthulhu_api_manager_rest.SuccessResponse-rest_GetNodePodPIDMappingResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rest.GetNodePodPIDMappingResponse"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "github_com_Gthulhu_api_manager_rest.SuccessResponse-rest_ListScheduleIntentsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rest.ListScheduleIntentsResponse"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "github_com_Gthulhu_api_manager_rest.SuccessResponse-rest_ListSchedulerStrategiesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rest.ListSchedulerStrategiesResponse"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "rest.CreateScheduleStrategyRequest": {
            "type": "object",
            "properties": {
                "commandRegex": {
                    "type": "string"
                },
                "executionTime": {
                    "type": "integer"
                },
                "k8sNamespace": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "labelSelectors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.LabelSelector"
                    }
                },
                "priority": {
                    "type": "integer"
                },
                "strategyNamespace": {
                    "type": "string"
                }
            }
        },
...
        "rest.DeleteScheduleIntentsRequest": {
            "type": "object",
            "properties": {
                "intentIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "rest.DeleteScheduleStrategyRequest": {
            "type": "object",
            "properties": {
                "strategyId": {
                    "type": "string"
                }
            }
        },
        "rest.GetNodePodPIDMappingResponse": {
            "type": "object",
            "properties": {
                "node_id": {
                    "type": "string"
                },
                "node_name": {
                    "type": "string"
                },
                "pods": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.PodPIDInfo"
                    }
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "rest.ListScheduleIntentsResponse": {
            "type": "object",
            "properties": {
                "intents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.ScheduleIntent"
                    }
                }
            }
        },
        "rest.ListSchedulerStrategiesResponse": {
            "type": "object",
            "properties": {
                "strategies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.ScheduleStrategy"
                    }
                }
            }
        },
...
        "rest.PodPIDInfo": {
            "type": "object",
            "properties": {
                "pod_id": {
                    "type": "string"
                },
                "pod_uid": {
                    "type": "string"
                },
                "processes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.PodPIDProcess"
                    }
                }
            }
        },
...
        "rest.ScheduleIntent": {
            "type": "object",
            "properties": {
                "commandRegex": {
                    "type": "string"
                },
                "executionTime": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "k8sNamespace": {
                    "type": "string"
                },
                "nodeID": {
                    "type": "string"
                },
                "podID": {
                    "type": "string"
                },
                "podLabels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "priority": {
                    "type": "integer"
                },
                "state": {
                    "$ref": "#/definitions/domain.IntentState"
                },
                "strategyID": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}